Rename JSON Tags
- [+] refactor(storage.go): rename json tags for Info struct fields
- [+] Rename `context_key` to `contextkey`
- [+] Rename `cookie_value` to `cookie`
- [+] Rename `expiration_time` to `expiration`
- [+] Rename `qr_code_data` to `qrcodedata`

Note: This renaming is done to avoid confusion because the Info struct is used by the Fiber 2FA middleware and stored in a separate database.  Since The Fiber 2FA middleware requires an independent storage to maintain the 2FA-related data, which is why it uses c.Locals to allow set values from another database (e.g., the main database) for the 2FA functionality.